backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic>=2.6.4
pydantic[email]>=2.6.4
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
cryptography>=42.0.8
bcrypt>=4.0.1